[voicestreams] converted decoding to calc as int16s and not floats

+++ libs/voicestreams/voicestreams-decode.h
@@ -26,7 +26,7 @@
 #define VOICESTREAMS_DECODED_AMBE_FRAME_SAMPLES_COUNT 160
 
 typedef struct {
-	float samples[VOICESTREAMS_DECODED_AMBE_FRAME_SAMPLES_COUNT];
+	int16_t samples[VOICESTREAMS_DECODED_AMBE_FRAME_SAMPLES_COUNT];
 } voicestreams_decoded_frame_t;
 
 voicestreams_decoded_frame_t *voicestreams_decode_ambe_frame(dmrpacket_payload_ambe_frame_bits_t *ambe_frame_bits, voicestream_t *voicestream);
